requirements.txt
fastapi
uvicorn[standard]
sympy
requests
python-dotenv
